Added character bot
This bot will be used to store all the values on a per player basis. The use of inline keyboards will make it easier to set specific values and keep it more constrained on what the user can store. Will also have a set command to change values, based on health, equipment, etc.